+++ pages/admin/settings/sso-connection/new.tsx
@@ -1,7 +1,7 @@
 import type { GetServerSidePropsContext, NextPage } from 'next';
 import CreateConnection from '@components/connection/CreateConnection';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-import type { AdminPortalSSODefaults } from '@components/connection/utils';
+import type { AdminPortalSSODefaults } from '@lib/utils';
 import { adminPortalSSODefaults } from '@lib/env';
 
 type Props = {
